refactor: extract daemon selector capture runtime
